+++ src/test/java/tests/SwitchToConstructorTest.java
@@ -2,7 +2,6 @@ package tests;
 
 import apiclient.UserService;
 import io.qameta.allure.Description;
-import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.filter.log.RequestLoggingFilter;
@@ -36,13 +35,11 @@ public class SwitchToConstructorTest {
     private static final String pageUrl = "https://stellarburgers.nomoreparties.site/";
 
     @BeforeClass
-    @Step("Глобальная настройка тестов")
     public static void globalSetUp() {
         RestAssured.filters(new RequestLoggingFilter(), new ResponseLoggingFilter());
     }
 
     @Before
-    @Step("Подготовка данных для теста")
     public void setUp() {
         driver = createWebDriver();
         userService = new UserService();
@@ -57,7 +54,6 @@ public class SwitchToConstructorTest {
     }
 
     @Test
-    @Step("Переход из личного кабинета в конструктор")
     @DisplayName("Тест перехода из личного кабинета в конструктор по клику на «Конструктор» и на логотип Stellar Burgers")
     @Description("Тест проверяет, что пользователь может успешно перейти из личного кабинета в конструктор по клику на «Конструктор» и на логотип Stellar Burgers")
     public void switchingToConstructorTest() {
@@ -78,10 +74,11 @@ public class SwitchToConstructorTest {
         assertEquals(expectedConstructorText, actualConstructorText);
     }
 
-    @Step("Очистка данных и завершение работы браузера после теста")
     @After
     public void tearDownAndClearData() {
         driver.quit();
-        userService.delete(token);
+        if (token != null) {
+            userService.delete(token);
+        }
     }
 }
